backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
alembic
pymysql

